Client: core design
