+++ packages/client-old/package.json
@@ -1,0 +1,68 @@
+{
+  "name": "@accounts/client-old",
+  "version": "0.1.0-beta.10",
+  "description": "Fullstack authentication and accounts-management",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "start": "webpack -p --config --progress --watch",
+    "clean": "rimraf lib",
+    "precompile": "npm run clean",
+    "compile": "tsc",
+    "prepublishOnly": "npm run compile",
+    "test": "npm run testonly",
+    "test-ci": "npm lint && npm coverage",
+    "testonly": "jest",
+    "test:watch": "jest --watch",
+    "coverage": "npm run testonly -- --coverage"
+  },
+  "jest": {
+    "transform": {
+      ".(ts|tsx)": "<rootDir>/../../node_modules/ts-jest/preprocessor.js"
+    },
+    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx)$",
+    "moduleFileExtensions": [
+      "ts",
+      "js"
+    ],
+    "testEnvironment": "jsdom",
+    "mapCoverage": true
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/js-accounts/accounts/tree/master/packages/client"
+  },
+  "keywords": [
+    "rest",
+    "graphql",
+    "grant",
+    "auth",
+    "authentication",
+    "accounts",
+    "users",
+    "oauth"
+  ],
+  "author": "Tim Mikeladze",
+  "license": "MIT",
+  "devDependencies": {
+    "@types/jest": "22.2.2",
+    "jest": "22.4.3",
+    "jsonwebtoken": "8.2.0",
+    "lerna": "2.9.1",
+    "localstorage-polyfill": "1.0.1",
+    "regenerator-runtime": "0.11.1",
+    "rimraf": "2.6.2"
+  },
+  "dependencies": {
+    "@accounts/common": "^0.1.0-beta.10",
+    "crypto-js": "^3.1.9-1",
+    "immutable": "^3.8.1",
+    "jwt-decode": "^2.1.0",
+    "lodash": "^4.16.4",
+    "redux": "^3.7.2",
+    "redux-immutable": "^4.0.0"
+  }
+}
