+++ packages/client-old/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../tsconfig",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./lib",
+    "lib": ["dom", "es6", "es2015", "es2016", "es2017"],
+    "typeRoots": [
+      "node_modules/@types"
+    ]
+  },
+  "exclude": [
+    "node_modules",
+    "__tests__",
+    "tmp_tests_",
+    "lib"
+  ]
+}
