+++ packages/client/__tests__/index.ts
@@ -1,0 +1,7 @@
+import AccountsClient from '../src';
+
+describe('AccountsClient entry', () => {
+  it('should have default export AccountsClient', () => {
+    expect(typeof AccountsClient).toBe('function');
+  });
+});
