+++ packages/client/src/index.ts
@@ -1,10 +1,1 @@
-import Accounts, { AccountsClient } from './accounts-client';
-import { TransportInterface } from './transport-interface';
-import config, { TokenStorage } from './config';
-import reducer from './module';
-
-export default Accounts;
-
-export { AccountsClient, config, reducer };
-
-export { TransportInterface, TokenStorage };
+export { default } from './accounts-client';
