+++ packages/client/src/types/configuration.ts
@@ -1,0 +1,10 @@
+export interface Configuration {
+
+	tokenStorage: any;
+
+	userStorage: any;
+
+	transport: any;
+	
+	strategies: any;
+}
