+++ packages/client/tsconfig.json
@@ -1,17 +1,12 @@
 {
-  "extends": "../../tsconfig",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "rootDir": "./src",
-    "outDir": "./lib",
-    "lib": ["dom", "es6", "es2015", "es2016", "es2017"],
-    "typeRoots": [
-      "node_modules/@types"
-    ]
+    "outDir": "./lib"
   },
   "exclude": [
     "node_modules",
     "__tests__",
-    "tmp_tests_",
     "lib"
   ]
 }
